Moved reshape.rs to /traits
